Rationalise default mod creation Update messaging Fix NRE for mod commands
