+++ pkg/modinstaller/install.go
@@ -2,7 +2,6 @@ package modinstaller
 
 import (
 	"context"
-
 	"github.com/turbot/go-kit/helpers"
 	"github.com/turbot/steampipe/pkg/error_helpers"
 	"github.com/turbot/steampipe/pkg/utils"
@@ -18,7 +17,7 @@ func InstallWorkspaceDependencies(ctx context.Context, opts *InstallOpts) (_ *In
 	}()
 
 	// install workspace dependencies
-	installer, err := NewModInstaller(ctx, opts)
+	installer, err := NewModInstaller(opts)
 	if err != nil {
 		return nil, err
 	}
