+++ pkg/modinstaller/uninstall.go
@@ -18,7 +18,7 @@ func UninstallWorkspaceDependencies(ctx context.Context, opts *InstallOpts) (*In
 	}()
 
 	// uninstall workspace dependencies
-	installer, err := NewModInstaller(ctx, opts)
+	installer, err := NewModInstaller(opts)
 	if err != nil {
 		return nil, err
 	}
