+++ pkg/modinstaller/verify_location.go
@@ -22,7 +22,7 @@ func ValidateModLocation(ctx context.Context, workspacePath string) bool {
 
 	// check if running in home directory
 	if workspacePath == home {
-		return utils.UserConfirmation(fmt.Sprintf("%s: You're in the home directory. It's recommended to create a new directory and run %s from there.\nDo you want to continue? (y/n)", color.YellowString("Warning"), constants.Bold(fmt.Sprintf("steampipe mod %s", cmd.Name()))))
+		return utils.UserConfirmation(fmt.Sprintf("%s: Creating a mod file in the home directory is not recommended.\nBest practice is to create a new directory and run %s from there.\nDo you want to continue? (y/n)", color.YellowString("Warning"), constants.Bold(fmt.Sprintf("steampipe mod %s", cmd.Name()))))
 	}
 	// else check if running in a directory containing lot of sql and sp files
 	fileList, _ := filehelpers.ListFiles(workspacePath, &filehelpers.ListOptions{
@@ -31,7 +31,7 @@ func ValidateModLocation(ctx context.Context, workspacePath string) bool {
 		MaxResults: MaxResults,
 	})
 	if len(fileList) == MaxResults {
-		return utils.UserConfirmation(fmt.Sprintf("%s: You're in a directory with a lot of files or subdirectories (>10 files that are not .sql or .sp). It's recommended to create a new directory and run %s from there.\nDo you want to continue? (y/n)", color.YellowString("Warning"), constants.Bold(fmt.Sprintf("steampipe mod %s", cmd.Name()))))
+		return utils.UserConfirmation(fmt.Sprintf("%s: Creating a mod file in a directory with a lot of files or subdirectories is not recommended.\nBest practice is to create a new directory and run %s from there.\nDo you want to continue? (y/n)", color.YellowString("Warning"), constants.Bold(fmt.Sprintf("steampipe mod %s", cmd.Name()))))
 	}
 
 	return true
